remove requirement.txt move models.py
