--- requirements.txt
@@ -1,5 +1,0 @@
-numpy==1.18.*
-nltk==3.5.*
-fastapi[all]==0.61.* 
-python-dotenv==0.14.*
-pymongo==3.11.*
