+++ router/user.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter
 import config
 import pymongo 
-import models
+from server import models
 import logging
 
 
@@ -27,8 +27,8 @@ async def update(user_stats:models.UserStats):
     #user = UserStats(user_name=user_name, score=score, num_games=num_games)
         
 
-    filer = {'user_name': user_stats.user_name}
-    res:pymongo.results.UpdateResult = db['user'].replace_one(filer ,user_stats.dict() , upsert= True)
+    query_filter = {'user_name': user_stats.user_name}
+    res:pymongo.results.UpdateResult = db['user'].replace_one(query_filter ,user_stats.dict() , upsert= True)
     
     logging.debug(f"Updated {user_stats.user_name} with {user_stats}")
     return res.acknowledged
